ECR-2330 Remove old map proofs
