+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/map/CheckedMapProof.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.exonum.binding.common.proofs.map.flat;
+package com.exonum.binding.common.proofs.map;
 
 import com.exonum.binding.common.hash.HashCode;
 import java.util.List;
