--- exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/map/LeafMapProofNode.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2018 The Exonum Team
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.exonum.binding.common.proofs.map;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
-/**
- * A proof node for a map that contains a mapping for the requested key.
- */
-public final class LeafMapProofNode implements MapProofNode {
-
-  private final byte[] value;
-
-  public LeafMapProofNode(byte[] value) {
-    this.value = checkNotNull(value);
-  }
-
-  public byte[] getValue() {
-    return value;
-  }
-
-  @Override
-  public void accept(MapProofVisitor visitor) {
-    visitor.visit(this);
-  }
-}
