+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/map/MapEntry.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.exonum.binding.common.proofs.map.flat;
+package com.exonum.binding.common.proofs.map;
 
 import java.util.Arrays;
