+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/map/UncheckedFlatMapProof.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.exonum.binding.common.proofs.map.flat;
+package com.exonum.binding.common.proofs.map;
 
 import static com.exonum.binding.common.hash.Funnels.hashCodeFunnel;
 import static com.exonum.binding.common.proofs.DbKeyFunnel.dbKeyFunnel;
@@ -22,7 +22,6 @@ import static com.exonum.binding.common.proofs.DbKeyFunnel.dbKeyFunnel;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.hash.HashFunction;
 import com.exonum.binding.common.hash.Hashing;
-import com.exonum.binding.common.proofs.map.DbKey;
 import com.exonum.binding.common.proofs.map.DbKey.Type;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
