+++ exonum-java-binding-common/src/test/java/com/exonum/binding/common/proofs/map/DbKeyTest.java
@@ -61,84 +61,6 @@ class DbKeyTest {
   }
 
   @Test
-  void throwsIfBranchKeySliceHasBitsAfterSignificantPart0Bits() {
-    if (MapProofValidator.PERFORM_TREE_CORRECTNESS_CHECKS) {
-      int numSignificantBits = 0;
-      byte[] rawDbKey = createDbKey(Type.BRANCH.code, bytes(0x01), numSignificantBits);
-
-      assertThrows(IllegalArgumentException.class, () -> {
-        DbKey dbKey = DbKey.fromBytes(rawDbKey);
-      });
-
-    }
-  }
-
-  @Test
-  void throwsIfBranchKeySliceHasBitsAfterSignificantPart1Bit() {
-    if (MapProofValidator.PERFORM_TREE_CORRECTNESS_CHECKS) {
-      int numSignificantBits = 1;
-      byte[] rawDbKey = createDbKey(Type.BRANCH.code, bytes(0x03), numSignificantBits);
-
-      assertThrows(IllegalArgumentException.class, () -> {
-        DbKey dbKey = DbKey.fromBytes(rawDbKey);
-      });
-
-    }
-  }
-
-  @Test
-  void throwsIfBranchKeySliceHasBitsAfterSignificantPart7Bit() {
-    if (MapProofValidator.PERFORM_TREE_CORRECTNESS_CHECKS) {
-      int numSignificantBits = 7;
-      byte[] rawDbKey = createDbKey(Type.BRANCH.code, bytes(0x80), numSignificantBits);
-
-      assertThrows(IllegalArgumentException.class, () -> {
-        DbKey dbKey = DbKey.fromBytes(rawDbKey);
-      });
-
-    }
-  }
-
-  @Test
-  void throwsIfBranchKeySliceHasBitsAfterSignificantPart8Bit() {
-    if (MapProofValidator.PERFORM_TREE_CORRECTNESS_CHECKS) {
-      int numSignificantBits = 8;
-      byte[] rawDbKey = createDbKey(Type.BRANCH.code, bytes(0xFF, 0x01), numSignificantBits);
-
-      assertThrows(IllegalArgumentException.class, () -> {
-        DbKey dbKey = DbKey.fromBytes(rawDbKey);
-      });
-
-    }
-  }
-
-  @Test
-  void throwsIfBranchKeySliceHasBitsAfterSignificantPart12Bit() {
-    if (MapProofValidator.PERFORM_TREE_CORRECTNESS_CHECKS) {
-      int numSignificantBits = 12;
-      byte[] rawDbKey = createDbKey(Type.BRANCH.code, bytes(0xFF, 0x1F), numSignificantBits);
-
-      assertThrows(IllegalArgumentException.class, () -> {
-        DbKey dbKey = DbKey.fromBytes(rawDbKey);
-      });
-
-    }
-  }
-
-  @Test
-  void throwsIfBranchKeySliceHasBitsAfterSignificantPart16Bit() {
-    if (MapProofValidator.PERFORM_TREE_CORRECTNESS_CHECKS) {
-      int numSignificantBits = 16;
-      byte[] rawDbKey = createDbKey(Type.BRANCH.code, bytes(0xFF, 0xFF, 0x01), numSignificantBits);
-
-      assertThrows(IllegalArgumentException.class, () -> {
-        DbKey dbKey = DbKey.fromBytes(rawDbKey);
-      });
-
-    }
-  }
-
-  @Test
   void getNodeType_BranchNode() {
     byte[] rawDbKey = createDbKey(Type.BRANCH.code, bytes("a"), 8);
     DbKey dbKey = DbKey.fromBytes(rawDbKey);
