+++ exonum-java-binding-core/src/test/java/com/exonum/binding/storage/indices/MapProofValidatorMatcher.java
@@ -19,7 +19,6 @@ package com.exonum.binding.storage.indices;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.exonum.binding.common.hash.HashCode;
-import com.exonum.binding.common.proofs.map.MapProofValidator;
 import javax.annotation.Nullable;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
